requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
sentence-transformers
faiss-cpu
openai
tqdm
langchain
transformers

